pyproject.toml
[tool.poetry]
name = "qa-test-generator"
version = "0.1.0"
description = "AI-powered QA Test Case Generator from Figma designs and requirement documents"
authors = ["QA Automation Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
httpx = "^0.25.2"
python-multipart = "^0.0.6"
pydantic = "^2.5.2"
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.23"
alembic = "^1.13.0"
asyncpg = "^0.29.0"
redis = "^5.0.1"
python-docx = "^1.1.0"
pypdf2 = "^3.0.1"
anthropic = "^0.8.1"
python-dotenv = "^1.0.0"
aiofiles = "^23.2.1"
tenacity = "^8.2.3"
structlog = "^23.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.12.0"
ruff = "^0.1.8"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
